// src/shared/enums/feature.enum.ts
export enum FeaturesEnum {
  DEPARTMENT_LIST = "DEPARTMENT:LIST",
  DEPARTMENT_CREATE = "DEPARTMENT:CREATE",
  DEPARTMENT_UPDATE = "DEPARTMENT:UPDATE",
  DEPARTMENT_DELETE = "DEPARTMENT:DELETE",

  ESTABLISHMENT_LIST = "ESTABLISHMENT:LIST",
  ESTABLISHMENT_CREATE = "ESTABLISHMENT:CREATE",
  ESTABLISHMENT_UPDATE = "ESTABLISHMENT:UPDATE",
  ESTABLISHMENT_DETAIL = "ESTABLISHMENT:DETAIL",

  INSTANCE_LIST = "INSTANCE:LIST",
  INSTANCE_DETAIL = "INSTANCE:DETAIL",
  INSTANCE_CREATE = "INSTANCE:CREATE",
  INSTANCE_UPDATE = "INSTANCE:UPDATE",
  INSTANCE_DELETE = "INSTANCE:DELETE",

  PHARMACIST_LIST = "PHARMACIST:LIST",
  PHARMACIST_CREATE = "PHARMACIST:CREATE",
  PHARMACIST_UPDATE = "PHARMACIST:UPDATE",
  PHARMACIST_DELETE = "PHARMACIST:DELETE",
  PHARMACIST_DETAIL = "PHARMACIST:DETAIL",

  PROFILE_LIST = "PROFILE:LIST",
  PROFILE_DETAIL = "PROFILE:DETAIL",
  PROFILE_CREATE = "PROFILE:CREATE",
  PROFILE_UPDATE = "PROFILE:UPDATE",
  PROFILE_DELETE = "PROFILE:DELETE",

  FEATURE_LIST = "FEATURE:LIST",
  FEATURE_DETAIL = "FEATURE:DETAIL",

  ROLES_LIST = "ROLES:LIST",
  ROLES_DETAIL = "ROLES:DETAIL",
  ROLES_CREATE = "ROLES:CREATE",
  ROLES_UPDATE = "ROLES:UPDATE",
  ROLES_ASSIGN = "ROLES:ASSIGN",
  ROLES_REQUEST = "ROLES:REQUEST",
  ROLES_MODERATE = "ROLES:MODERATE",

  USER_LIST = "USER:LIST",
  USER_DETAIL = "USER:DETAIL",

  ADDRESS_CREATE = "ADDRESS:CREATE",
  ADDRESS_LIST = "ADDRESS:LIST",
  ADDRESS_DETAIL = "ADDRESS:DETAIL",
  ADDRESS_UPDATE = "ADDRESS:UPDATE",

  DOMAIN_LIST = "DOMAIN:LIST",
  PRODUCT_LIST = "PRODUCT:LIST",

  ENTRY_CREATE = "ENTRY:CREATE",
  ENTRY_LIST = "ENTRY:LIST",
  ENTRY_UPDATE = "ENTRY:UPDATE",
  ENTRY_DELETE = "ENTRY:DELETE",

  OUTBOUND_CREATE = "OUTBOUND:CREATE",
  OUTBOUND_LIST = "OUTBOUND:LIST",
  OUTBOUND_UPDATE = "OUTBOUND:UPDATE",
  OUTBOUND_DELETE = "OUTBOUND:DELETE",

  HEALTHUSER_CREATE = "HEALTHUSER:CREATE",
  HEALTHUSER_LIST = "HEALTHUSER:LIST",
  HEALTHUSER_UPDATE = "HEALTHUSER:UPDATE",
  HEALTHUSER_DELETE = "HEALTHUSER:DELETE",

  STOCKPOSITION_LIST = "STOCKPOSITION:LIST",

  DISPENSATION_LIST = "DISPENSATION:LIST",
  DISPENSATION_CREATE = "DISPENSATION:CREATE",
  DISPENSATION_UPDATE = "DISPENSATION:UPDATE",
  DISPENSATION_DELETE = "DISPENSATION:DELETE",
  DISPENSATION_DETAIL = "DISPENSATION:DETAIL",
  DISPENSATION_REVERSE = "DISPENSATION:REVERSE",
  OPEN = "OPEN",

  WORKSPACE_LIST = "WORKSPACE:LIST",
  WORKSPACE_CREATE = "WORKSPACE:CREATE",
  WORKSPACE_UPDATE = "WORKSPACE:UPDATE",
  WORKSPACE_DELETE = "WORKSPACE:DELETE",
  WORKSPACE_INVITE = "WORKSPACE:INVITE",
}